refactor: 사다리 생성 방식 변경
