+++ src/main/java/nextstep/ladder/domain/LadderGame.java
@@ -1,27 +1,24 @@
 package nextstep.ladder.domain;
 
-import nextstep.ladder.domain.player.Player;
+import nextstep.ladder.domain.ladder.Ladder;
+import nextstep.ladder.domain.player.Players;
 import nextstep.ladder.dto.LadderDto;
 import nextstep.ladder.dto.PlayerDto;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class LadderGame {
 
-    private final List<Player> players;
+    private final Players players;
     private final Ladder ladder;
 
-    public LadderGame(List<String> playerNames, int ladderHeight) {
-        this.players = playerNames.stream()
-                                  .map(Player::new)
-                                  .collect(Collectors.toList());
-
-        ladder = Ladder.of(ladderHeight, players.size());
+    public LadderGame(Players players, Ladder ladder) {
+        this.players = players;
+        this.ladder = ladder;
     }
 
     public List<PlayerDto> getPlayers() {
-        return players.stream().map(Player::export).collect(Collectors.toList());
+        return players.export();
     }
 
     public LadderDto getLadder() {
