+++ src/test/java/nextstep/ladder/domain/ladder/LadderTest.java
@@ -1,0 +1,5 @@
+package nextstep.ladder.domain.ladder;
+
+class LadderTest {
+
+}
